perf(tests): enable parallel test execution with CODEMIE_HOME isolation
- Add CODEMIE_HOME environment variable support (follows CARGO_HOME pattern)
- Create getCodemieHome() utility for configurable paths
- Implement setupTestIsolation() helper for automatic test isolation
- Split cli-commands.test.ts into 7 separate files for parallelization
- Update all hardcoded ~/.codemie paths to use getCodemieHome()
- Configure Vitest thread pool (8 threads, isolation enabled)

Performance: Reduced integration test time from 12.9s to 6.53s (2x faster)
Test files can now run in parallel without state conflicts

Files changed:
- src/utils/codemie-home.ts (NEW)
- tests/helpers/test-isolation.ts (NEW)
- tests/integration/cli-commands/*.test.ts (7 NEW files)
- vitest.config.ts (parallel execution enabled)
- All path consumers updated (config-loader, credential-store, metrics, etc.)
